Skip over empty strings when inferring a containers DNS name.
When a container is launched with "DNSDOCK_IMAGE=", we should ignore
then empty image name so that we can create an appropriate DNS name
for the container.

Also riding on this patch is a change to how queryServices determines
whether a service name should be returned for a given query. I've
refactored the code to a single function called isPrefixQuery()
and included docstring/tests for it. This should greatly simplify
this codepath.