Support for Containers w/ Uppercase Characters
PROBLEM SUMMARY:

If a container named "bigTEST" was created while dnsdock was running,
using "dig @172.17.42.1 bigTEST.test_image.docker" would NOT resolve.
However, if the container was named "bigtest", it would work just fine.
Handling uppercase letters in the service name is was the crux of the
problem.

SOLUTION:

Because users in Docker are free to use uppercase letters for naming the
service, DNS should resolve the container name correctly because DNS is
case-insenstive. Hence "bigTEST.test_image.docker" should be equivalent
to "bigtest.test_image.docker".

A problem not solved here is if a user starts "bigTEST" and "bigtest".
Then you end up with multiple DNS A records ... which may not be correct
behaviour.