Allow tag com.dnsdock.prefix to select which of a container's IPs is added
For containers that are both on a private, internal network and a
bridge network, dnsdock often chooses the address of the internal
network which is not accessible from the host. Setting the prefix
to the address prefix the bridge network uses allows to prevent this.

Usage: -l com.dnsdock.prefix=172.17.0.